requirements.txt
colour
descartes
ete3
geopandas
shapely
matplotlib
pandas
pyyaml
requests
tabulate
unidecode
wget
